+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### Change Log
 
 #### 3.1.11 (To be released)
+  Bugs
+  * Fix #1013 : Kubernetes connection is not getting closed.
 
   Bugs
    * Impersonation parameters not set in withRequestConfig - https://github.com/fabric8io/kubernetes-client/pull/1037